lyrics_service/requirements.txt
uvicorn
fastapi
google.generativeai